Refactoring game-rule
