+++ plugins/game-rule-dnd5e/src/main.tsx
@@ -11,13 +11,15 @@
 // These imports are crucial for TypeScript type checking but are removed during the build process
 // to prevent bundling these libraries within the plugin, thus avoiding the "two Reacts" problem.
 import type { WritableDraft } from "immer";
-import type { Plugin, PluginWrapper, StoredState } from "@/lib/state";
+import type { Plugin, PluginWrapper, StoredState, IGameRuleLogic, Character } from "@/lib/state";
 import type { Context } from "@/app/plugins";
 import type * as Immer from 'immer';
 import type * as RadixThemes from '@radix-ui/themes';
 import type { ChangeEvent } from 'react';
 import type { useShallow } from 'zustand/shallow';
-import * as z from "zod";
+import type { Prompt } from "@/lib/prompts";
+import { DndStatsSettings, generateDefaultDndStatsSettings, DndClassData, DND_CLASS_DATA, DndStatsSettingsSchema } from "./pluginData";
+import { getProtagonistGenerationPrompt } from "./pluginPrompt";
 
 
 // Declare a module-level React variable.
@@ -25,153 +27,6 @@ import * as z from "zod";
 // This is essential for JSX transformation and ensuring all React operations within the plugin
 // use the same React instance as the main application, preventing "two Reacts" issues.
 let React: typeof import('react');
-
-/**
- * Zod schema for validating D&D character stats settings.
- * Each stat is an integer between 1 and 20.
- */
-const DndStatsSettingsSchema = z.object({
-  strength: z.number().int().min(1).max(20),
-  dexterity: z.number().int().min(1).max(20),
-  constitution: z.number().int().min(1).max(20),
-  intelligence: z.number().int().min(1).max(20),
-  wisdom: z.number().int().min(1).max(20),
-  charisma: z.number().int().min(1).max(20),
-  dndClass: z.string(),
-  dndSubclass: z.string(),
-});
-
-/**
- * TypeScript type inferred from the DndStatsSettingsSchema.
- * Represents the structure of the D&D character stats.
- */
-type DndStatsSettings = z.infer<typeof DndStatsSettingsSchema>;
-
-/**
- * Function to generate default values for D&D character stats using dice rolls.
- * Used when no settings are provided or to fill in missing values.
- */
-const generateDefaultDndStatsSettings = (rpgDiceRoller: typeof import('@dice-roller/rpg-dice-roller')): DndStatsSettings => {
-  const rollFormula = "4d6dl1"; // 4 six-sided dice, drop the lowest
-
-  const rollAttribute = () => new rpgDiceRoller.DiceRoll(rollFormula).total;
-
-  const generated = {
-    strength: rollAttribute(),
-    dexterity: rollAttribute(),
-    constitution: rollAttribute(),
-    intelligence: rollAttribute(),
-    wisdom: rollAttribute(),
-    charisma: rollAttribute(),
-    dndClass: "",
-    dndSubclass: "",
-  };
-  //console.log("generateDefaultDndStatsSettings generated:", generated);
-  return generated;
-};
-
-type DndClassData = {
-  [key: string]: string[];
-};
-
-const DND_CLASS_DATA: DndClassData = {
-  "Barbarian": [
-    "Path of the Berserker",
-    "Path of the Totem Warrior",
-    "Path of the Ancestral Guardian",
-    "Path of the Storm Herald",
-    "Path of the Zealot"
-  ],
-  "Bard": [
-    "College of Lore",
-    "College of Valor",
-    "College of Glamour",
-    "College of Whispers"
-  ],
-  "Cleric": [
-    "Life Domain",
-    "Light Domain",
-    "Trickery Domain",
-    "Knowledge Domain",
-    "Nature Domain",
-    "Tempest Domain",
-    "War Domain",
-    "Death Domain",
-    "Forge Domain",
-    "Grave Domain"
-  ],
-  "Druid": [
-    "Circle of the Moon",
-    "Circle of the Land",
-    "Circle of Dreams",
-    "Circle of the Shepherd"
-  ],
-  "Fighter": [
-    "Champion",
-    "Battle Master",
-    "Eldritch Knight",
-    "Arcane Archer",
-    "Cavalier",
-    "Samurai"
-  ],
-  "Monk": [
-    "Way of the Open Hand",
-    "Way of the Shadow",
-    "Way of the Four Elements",
-    "Way of the Drunken Master",
-    "Way of the Kensei",
-    "Way of the Sun Soul"
-  ],
-  "Paladin": [
-    "Oath of Devotion",
-    "Oath of the Ancients",
-    "Oath of Vengeance",
-    "Oathbreaker",
-    "Oath of Conquest",
-    "Oath of Redemption"
-  ],
-  "Ranger": [
-    "Hunter",
-    "Beast Master",
-    "Gloom Stalker",
-    "Horizon Walker",
-    "Monster Slayer"
-  ],
-  "Rogue": [
-    "Thief",
-    "Assassin",
-    "Arcane Trickster",
-    "Inquisitive",
-    "Mastermind",
-    "Scout",
-    "Swashbuckler"
-  ],
-  "Sorcerer": [
-    "Draconic Bloodline",
-    "Wild Magic",
-    "Divine Soul",
-    "Shadow Magic",
-    "Storm Sorcery"
-  ],
-  "Warlock": [
-    "The Archfey",
-    "The Fiend",
-    "The Great Old One",
-    "The Celestial",
-    "The Hexblade"
-  ],
-  "Wizard": [
-    "School of Abjuration",
-    "School of Conjuration",
-    "School of Divination",
-    "School of Enchantment",
-    "School of Evocation",
-    "School of Illusion",
-    "School of Necromancy",
-    "School of Transmutation",
-    "War Magic"
-  ]
-};
 
 /**
  * Props for the DndStatsCharacterUIPage component.
@@ -296,7 +151,7 @@ const DndStatsCharacterUIPage = ({
  * Main plugin class for D&D Character Stats.
  * Implements the Plugin interface to integrate with the Waidrin application.
  */
-export default class DndStatsPlugin implements Plugin {
+export default class DndStatsPlugin implements Plugin, IGameRuleLogic {
   private context: Context | undefined; // The plugin's context provided by the application
   private settings: DndStatsSettings | undefined; // The plugin's settings
 
@@ -341,5 +196,21 @@ export default class DndStatsPlugin implements Plugin {
         }}
       />
     );
+  }
+
+  getGameRuleLogic(): IGameRuleLogic {
+    return this;
+  }
+
+  async getInitialProtagonistStats(): Promise<string> {
+    if (!this.settings || !this.context) {
+        return "";
+    }
+    const stats = this.settings as DndStatsSettings;
+    const pc = this.context.getGlobalState();
+    const prompt = getProtagonistGenerationPrompt(stats, pc);
+
+    const narration = await this.context!.getBackend().getNarration(prompt);
+    return narration;
   }
 }
